% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsConversionPlot.R
\name{bsConversionPlot}
\alias{bsConversionPlot}
\title{Mehtylation bias plot}
\usage{
bsConversionPlot(bs)
}
\arguments{
\item{Takes}{the bs object as the input}
}
\value{
Plot showing bisulfite conversion rate for each sample
}
\description{
Plot the bisulfite conversion rate for each sample
based on the pheno data in the bs object
}
\examples{
bsConversionPlott(bs)
}
