% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromosomeCoverage.R
\name{chromosomeCoverage}
\alias{chromosomeCoverage}
\title{square a number}
\usage{
chromosomeCoverage(bs)
}
\arguments{
\item{x}{A numeric input}
}
\value{
The sqaure of the output
}
\description{
Take in a numeric value and squares it
}

