% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combineRDA.R
\name{combineRDA}
\alias{combineRDA}
\title{Combine multiple bsseq rda object}
\usage{
combineRDA(rdaList)
}
\arguments{
\item{rdaList}{list of bsseq rda objects}
}
\value{
Combined one rda object for all the cells
}
\description{
Combine multiple bsseq rda object into one bs object
Preferably for a library pool of cells
}
\examples{
bsObject1<-createRDA(system.file("extdata",'sc-RRBS_zyg_01_chr1_CpG.bedGraph',package='scmeth'))
bsObject2<-createRDA(system.file("extdata",'sc-RRBS_zyg_02_chr1_CpG.bedGraph',package='scmeth'))
combineRDA(c(bsObject1, bsObject2))
}
