% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage.R
\name{coverage}
\alias{coverage}
\title{Coverage statistics}
\usage{
coverage(bsObject)
}
\arguments{
\item{takes}{bs object as input}
}
\value{
vector of numbers as output
}
\description{
Provides Coverage metrics for the sample
}
\examples{
coverage(bsseqObject)
}
