% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpgDiscretization.R
\name{cpgDiscretization}
\alias{cpgDiscretization}
\title{Discretize the CpG methylation values
to align with single cell analysis}
\usage{
cpgDiscretization(bs)
}
\arguments{
\item{rda}{as an input}
}
\value{
discretized methylation matrix and a plot showing how many of the CpGs are discarded
}
\description{
Take in a numeric value and squares it
}
