% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createRDA.R
\name{createRDA}
\alias{createRDA}
\title{Generate bsseq rda object}
\usage{
createRDA(file)
}
\arguments{
\item{file}{CpG.bedGraph file output from methyldackel/pileometh}
}
\value{
bsseq object with coverga and methylation information for the sample
}
\description{
Generate bsseq rda object from the methyldackel/pileometh output
}
\examples{
createRDA(CpG.bedGraph file)
}
