% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downsample.R
\name{downsample}
\alias{downsample}
\title{Downsample metrics}
\usage{
downsample(bs, dsRates = c(0.01, 0.02, 0.05, seq(0.1, 0.9, 0.1), 0.99, 1))
}
\arguments{
\item{takes}{bs object as input}
}
\value{
Data frame with the saturation Data
}
\description{
Downsample the CpG covergae matrix for saturation analysis
}
\examples{
downsample(bsObject,c(0,0.5,1))
}
