% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbiasPlot.R
\name{mbiasplot}
\alias{mbiasplot}
\title{Methylation bias plot}
\usage{
mbiasplot(mbiasFile)
}
\arguments{
\item{mbiasFile}{methylation bias file provided by methyldackel/pileometh}
}
\value{
Plot showing the methylation across the read position in original top
and original bottom strand both in forward and reverse reads
}
\description{
Plot the methylation at each position of the read to observe any biases in the methylation
based on the read position
}
\examples{
mbiasplot(mbiasFile)
}
