% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methylationDist.R
\name{methylationDist}
\alias{methylationDist}
\title{Provide graphics for methylation distribution}
\usage{
methylationDist(bs)
}
\arguments{
\item{bs}{bsseq object}
}
\value{
plot of the methylation distribution
}
\description{
Plot the methylation distribution for the cells in bsseq object
}
\examples{
directory<-system.file("extdata/bismark_data",package='scmeth')
bs<-SummarizedExperiment::loadHDF5SummarizedExperiment(directory)
methylationDist(bs)
}
