% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readmetrics.R
\name{readmetrics}
\alias{readmetrics}
\title{Provide graphics for read information}
\usage{
readmetrics(readData)
}
\arguments{
\item{readData}{a .txt file providing mapped and unmapped reads for each sample}
}
\value{
Plot showing the mapped and unmapped read information for each cell
}
\description{
Plot the mapped and unmapped reads
}
\examples{
readmetrics(read.txt)
}
