% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repMask.R
\name{repMask}
\alias{repMask}
\title{Repeat masker coverage statistics}
\usage{
repMask(bs, organism, genome)
}
\arguments{
\item{Takes}{bs object, name of the organism and reference genome}
}
\value{
Data frame with sample name and coverage in repeat masker regions
}
\description{
Provides Coverage metrics based on the repeat masker region
}

