% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{report}
\alias{report}
\title{square a number}
\usage{
report(bsObj, outdirectory, organism, genome, meta = NULL, cacheable = NA)
}
\arguments{
\item{x}{A numeric input}
}
\value{
The sqaure of the output
}
\description{
Take in a numeric value and squares it
}

