% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scmeth.R
\docType{package}
\name{scmeth}
\alias{scmeth}
\alias{scmeth-package}
\title{scmeth: a package to conduct quality control analysis for methylation data.
Most functions can be applied to both bulk and single-cell methylation
while other functions are specific to single-cell methylation data.
scmeth is especially customized to use the output from the FireCloud
implementation of methylation pipeline to produce comprehensive
quality control report}
\description{
scmeth: a package to conduct quality control analysis for methylation data.
Most functions can be applied to both bulk and single-cell methylation
while other functions are specific to single-cell methylation data.
scmeth is especially customized to use the output from the FireCloud
implementation of methylation pipeline to produce comprehensive
quality control report
}
