% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsConversionPlot.R
\name{bsConversionPlot}
\alias{bsConversionPlot}
\title{Bisulfite conversion rate visualization}
\usage{
bsConversionPlot(bs)
}
\arguments{
\item{dir}{directory that contains bismark output}
}
\value{
Plot showing bisulfite conversion rate for each sample
}
\description{
Plot the bisulfite conversion rate for each sample
based on the pheno data in the bs object
}
\examples{
directory<-system.file("extdata/bismark_data",package='scmeth')
bs<-SummarizedExperiment::loadHDF5SummarizedExperiment(directory)
bsConversionPlot(bs)
}
