% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromosomeCoverage.R
\name{chromosomeCoverage}
\alias{chromosomeCoverage}
\title{Coverage statistics based on chromosome}
\usage{
chromosomeCoverage(bs)
}
\arguments{
\item{takes}{bs object as input}
}
\value{
matrix of numbers as output with
column and rows indicating the samples and the chromosome
}
\description{
Provides Coverage metrics for the sample by each chromosome
}
