% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromosomeCoverage.R
\name{chromosomeCoverage}
\alias{chromosomeCoverage}
\title{CpG covergae in each chromosome}
\usage{
chromosomeCoverage(bs)
}
\arguments{
\item{bs}{bsseq object}
}
\value{
matrix of chromsome covergae with
column and rows indicating the samples and the chromosome respectively
}
\description{
Provides Coverage metrics for each sample by the chromosome
}
\examples{
directory<-system.file("extdata/bismark_data",package='scmeth')
bs<-HDF5Array::loadHDF5SummarizedExperiment(directory)
chromosomeCoverage(bs)
}
