% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combineRDA.R
\name{combineRDA}
\alias{combineRDA}
\title{Combine multiple bsseq rda object}
\usage{
combineRDA(rdaList)
}
\arguments{
\item{rdaList}{list of bsseq rda objects}
}
\value{
Combined one rda object for all the cells
}
\description{
Combine multiple bsseq rda object into one bs object
Preferably for a library pool of cells
}
\examples{
CpGBedGraphFile_1<-system.file("extdata",'sc-RRBS_zyg_01_chr1_CpG.bedGraph',package='scmeth')
readMetricsFile_1<-system.file("extdata",'sc-RRBS-zygote_01.read_metrics.txt',package='scmeth')
bsConversionFile_1<-system.file("extdata",'sc-RRBS-zygote_01.bsConv.txt',package='scmeth')
CpGBedGraphFile_2<-system.file("extdata",'sc-RRBS_zyg_02_chr1_CpG.bedGraph',package='scmeth')
readMetricsFile_2<-system.file("extdata",'sc-RRBS-zygote_02.read_metrics.txt',package='scmeth')
bsConversionFile_2<-system.file("extdata",'sc-RRBS-zygote_02.bsConv.txt',package='scmeth')
bsObject1<-createRDA(CpGBedGraphFile_1,readMetricsFile_1,bsConversionFile_1)
bsObject2<-createRDA(CpGBedGraphFile_2,readMetricsFile_2,bsConversionFile_2)
combineRDA(c(bsObject1, bsObject2))
}
