% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combineRDA.R
\name{combineRDA}
\alias{combineRDA}
\title{Combine multiple bsseq rda object}
\usage{
combineRDA(rdaList)
}
\arguments{
\item{Takes}{bs object, name of the organism and reference genome}
}
\value{
Data frame with sample name and coverage in repeat masker regions
}
\description{
Combine multiple bsseq rda object into one bs object
Preferably for a library pool of cells
}
\examples{
combineRDA(c(bsObject1, bsObject2))
}
