% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage.R
\name{coverage}
\alias{coverage}
\title{square a number}
\usage{
coverage(bs)
}
\arguments{
\item{x}{A numeric input}
}
\value{
The sqaure of the output
}
\description{
Take in a numeric value and squares it
}

