% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage.R
\name{coverage}
\alias{coverage}
\title{Coverage for single cells}
\usage{
coverage(bs)
}
\arguments{
\item{bs}{bsseq object}
}
\value{
vector of coverage for the cells in bs object
}
\description{
Provides Coverage for each cell in a library pool
}
\examples{
directory<-system.file("extdata/bismark_data",package='scmeth')
bs<-SummarizedExperiment::loadHDF5SummarizedExperiment(directory)
coverage(bs)
}
