% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpgDensity.R
\name{cpgDensity}
\alias{cpgDensity}
\title{Provides Coverage metrics in the repeat masker region}
\usage{
cpgDensity(bs, organism, windowLength = 100)
}
\arguments{
\item{bs}{bsseq object}

\item{organism}{scientific name of the organism of interest, i.e. Mus musculus or Homo sapiens}

\item{genome}{reference alignment, i.e. mm10 or hg38}
}
\value{
Data frame with sample name and coverage in repeat masker regions
}
\description{
Provides Coverage metrics in the repeat masker region
}
\examples{
repMask(bsseqObject,'Mus musculus','mm10')
}
