% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpgDensity.R
\name{cpgDensity}
\alias{cpgDensity}
\title{Provides Coverage metrics in the repeat masker region}
\usage{
cpgDensity(bs, organism, windowLength = 1000)
}
\arguments{
\item{bs}{bsseq object}

\item{organism}{scientific name of the organism of interest,
i.e. Mus musculus or Homo sapiens}

\item{windowLength}{Length of the window to calculate the density}
}
\value{
Data frame with sample name and coverage in repeat masker regions
}
\description{
Provides Coverage metrics in the repeat masker region
}
\examples{
library(BSgenome.Hsapiens.NCBI.GRCh38)
directory<-system.file("extdata/bismark_data",package='scmeth')
bs<-SummarizedExperiment::loadHDF5SummarizedExperiment(directory)
cpgDensity(bs,Hsapiens,1000)
}
