% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createRDA.R
\name{createRDA}
\alias{createRDA}
\title{Generate bsseq rda object}
\usage{
createRDA(file)
}
\arguments{
\item{Takes}{bs object, name of the organism and reference genome}
}
\value{
Data frame with sample name and coverage in repeat masker regions
}
\description{
Generate bsseq rda object from the pileometh output
}
\examples{
createRDA(CpG.bedGraph file)
}
