% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createRDA.R
\name{createRDA}
\alias{createRDA}
\title{Generate bsseq rda object}
\usage{
createRDA(CpG_file, readmetric_file, bsConv_file)
}
\arguments{
\item{CpG_file}{CpG.bedGraph file output from methyldackel/pileometh}

\item{readmetric_file}{file containing read information}

\item{bsConv_file}{file containing bisulfite conversion information}
}
\value{
bsseq object with coverga and methylation information for the sample
}
\description{
Generate bsseq rda object from the methyldackel/pileometh output
}
\examples{
CpGBedGraphFile<-system.file("extdata",'sc-RRBS_zyg_01_chr1_CpG.bedGraph',
package='scmeth')
readMetricsFile<-system.file("extdata",'sc-RRBS-zygote_01.read_metrics.txt',
package='scmeth')
bsConversionFile<-system.file("extdata",'sc-RRBS-zygote_01.bsConv.txt',
package='scmeth')
createRDA(CpGBedGraphFile,readMetricsFile,bsConversionFile)
}
