% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downsample.R
\name{downsample}
\alias{downsample}
\title{Downsample analysis}
\usage{
downsample(bs, dsRates = c(0.01, 0.02, 0.05, seq(0.1, 0.9, 0.1)))
}
\arguments{
\item{bs}{bsseq object}

\item{dsRates}{downsampling rate. i.e. the probabaility of sampling
a single CpG
default is list of probabilities ranging from 0.01 to 1
This can be changed by the user, for more continuous saturation curve
dsRates can be changed to add more sampling rates}
}
\value{
Data frame with the CpG coverage for each sample at each
sampling rate
}
\description{
Downsample the CpG covergae matrix for saturation analysis
}
\examples{
directory<-system.file("extdata/bismark_data",package='scmeth')
bs<-SummarizedExperiment::loadHDF5SummarizedExperiment(directory)
downsample(bs)
downsample(bs,seq(0,1,length.out=20))

}
