% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featureCoverage.R
\name{featureCoverage}
\alias{featureCoverage}
\title{Coverage statistics based on the feature}
\usage{
featureCoverage(bs, features, genome)
}
\arguments{
\item{takes}{bs object and list of features as input}
}
\value{
matrix a data frame with feature names and th number of CpG covered
}
\description{
Provides Coverage metrics for the sample by each features provided by the user
}
