% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methylationDist.R
\name{methylationDist}
\alias{methylationDist}
\title{Provide graphics for methylation distribution}
\usage{
methylationDist(bs)
}
\arguments{
\item{bs}{bsseq object}
}
\value{
mean methylation for each sample
}
\description{
Plot the methylation distribution for the cells in bsseq object
}
\examples{
directory <- system.file("extdata/bismark_data", package='scmeth')
bs <- HDF5Array::loadHDF5SummarizedExperiment(directory)
methylationDist(bs)
}
