% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methylationDist.R
\name{methylationDist}
\alias{methylationDist}
\title{Mehtylation distribution function}
\usage{
methylationDist(bs, all = FALSE)
}
\arguments{
\item{Takes}{bs object, name of the organism and reference genome}
}
\value{
Data frame with sample name and coverage in repeat masker regions
}
\description{
Plot the methylation distribution for a few of the cells
}
\examples{
methylationDist(bsObject,all=FALSE)
}
