% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methylationDist.R
\name{methylationDist}
\alias{methylationDist}
\title{Provide graphics for methylation distribution}
\usage{
methylationDist(bs, all = FALSE)
}
\arguments{
\item{bs}{bsseq object}

\item{all}{Indicator of whether to plot the distribution for all the
cells or for
a single cell}
}
\value{
plot of the methylation distribution
}
\description{
Plot the methylation distribution for the cells in bsseq object
}
\examples{
load(system.file("extdata",'bsObject.rda',package='scmeth'))
methylationDist(bs)
methylationDist(bs,all = TRUE)
}
