% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methylationDist.R
\name{methylationDist}
\alias{methylationDist}
\title{Provide graphics for methylation distribution}
\usage{
methylationDist(bs, subSample = 1e+06, coverageVec = NULL)
}
\arguments{
\item{bs}{bsseq object}

\item{coverageVec}{If coverage vector is already calculated provide it to
speed up the process}
}
\value{
plot of the methylation distribution
}
\description{
Plot the methylation distribution for the cells in bsseq object
}
\examples{
directory<-system.file("extdata/bismark_data",package='scmeth')
bs<-SummarizedExperiment::loadHDF5SummarizedExperiment(directory)
methylationDist(bs)
}
