% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readmetrics.R
\name{readmetrics}
\alias{readmetrics}
\title{Provide graphics for read information}
\usage{
readmetrics(bs)
}
\arguments{
\item{bs}{bsseq object}
}
\value{
Plot showing the mapped and unmapped read information for each cell
}
\description{
Plot the mapped and unmapped reads
}
\examples{
directory <- system.file("extdata/bismark_data",package='scmeth')
bs <- HDF5Array::loadHDF5SummarizedExperiment(directory)
readmetrics(bs)
}
