% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readmetrics.R
\name{readmetrics}
\alias{readmetrics}
\title{Provide graphics for read information}
\usage{
readmetrics(bs)
}
\arguments{
\item{bs}{bsseq object}
}
\value{
Plot showing the mapped and unmapped read information for each cell
}
\description{
Plot the mapped and unmapped reads
}
\examples{
load(system.file("extdata",'bsObject.rda',package='scmeth'))
readmetrics(bs)
}
