% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repMask.R
\name{repMask}
\alias{repMask}
\title{square a number}
\usage{
repMask(bs, organism, genome)
}
\arguments{
\item{x}{A numeric input}
}
\value{
The sqaure of the output
}
\description{
Take in a numeric value and squares it
}

