% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repMask.R
\name{repMask}
\alias{repMask}
\title{Coverage of CpGs in repeat masker region}
\usage{
repMask(bs, organism, genome)
}
\arguments{
\item{bs}{bsseq object}

\item{organism}{scientific name of the organism of interest, i.e. Mus musculus or Homo sapiens}

\item{genome}{reference alignment, i.e. mm10 or hg38}
}
\value{
Data frame with sample name and coverage in repeat masker regions
#@examples
load(system.file("extdata",'bsObject.rda',package='scmeth'))
repMask(bs,'Mus musculus','mm10')
}
\description{
Provides Coverage metrics in the repeat masker region
}
