% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{report}
\alias{report}
\title{Generates an inclusive report on methylation analysis}
\usage{
report(bsObj, outdirectory, organism, genome, meta = NULL, cacheable = NA)
}
\arguments{
\item{list}{of inputs}
}
\value{
A report will be the output based on rmarkdown file
}
\description{
Takes bs object, output directory, name of the organism, reference genome
}
