% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{report}
\alias{report}
\title{Generates an inclusive report on methylation analysis}
\usage{
report(bsObj, outdirectory, organism, genome, readData = NULL)
}
\arguments{
\item{bsObj}{bsseq object}

\item{outdirectory}{name of the output directory where the report will be saved}

\item{organism}{scientific name of the organism of interest, i.e. Mus musculus or Homo sapiens}

\item{genome}{reference alignment, i.e. mm10 or hg38}

\item{readData}{read information file on the samples. This is an optional argument and if given
the report will have graphics on read information}
}
\value{
Report will be an html file
}
\description{
This function uses most of the functions in this package to generate a report for the user
}
\examples{
library(BSgenome.Mmusculus.UCSC.mm10)
load(system.file("extdata",'bsObject.rda',package='scmeth'))
report(bs,'~',Mmusculus,'mm10')
}
