% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{report}
\alias{report}
\title{Generates an inclusive report on methylation analysis}
\usage{
report(bsObj, outdirectory, organism, genome)
}
\arguments{
\item{bsObj}{bsseq object}

\item{outdirectory}{name of the output directory where the
report will be saved}

\item{organism}{scientific name of the organism of interest,
i.e. Mus musculus or Homo sapiens}

\item{genome}{reference alignment, i.e. mm10 or hg38
the report will have graphics on read information}
}
\value{
Report will be an html file
}
\description{
This function uses most of the functions in this package to
generate a report for the user
}
\examples{
library(BSgenome.Hsapiens.NCBI.GRCh38)
directory<-system.file("extdata/bismark_data",package='scmeth')
report(directory,'~',Hsapiens,'hg38')
}
